+++ Changelog.md
@@ -19,6 +19,7 @@ _aaaa.mm.dd_
 ### Changes
 * Remove redundant typecheck in `PrimitiveTypeArray::setDataAt()` in `NDEBUG` mode. A performance gain of ~25% was observed on the MNIST application (without a thorough profiling method).
 * tpgGraphDotImporter now has a line MAX_READ_SIZE limit of 4096 (instead of 1024), allowing to read greater tpg dot files.
+* Add a caching mechanism in `Data::PrimitiveTypeArray[2D]` to avoid reexecuting the `getAddressSpace()` regular expression more than once for each data type. (This had a huge overhead on execution time.)
 
 ### Bug fix
 * Adapt code for building GEGELATI with clang standard library: libc++. (see Issue #49 for mode details)